Clean up CordaX500Name
* Remove unused imports
* Move X500 name generation into test code as it's no longer used in the main code
