+++ experimental/src/main/kotlin/net/corda/finance/contracts/universal/PrettyPrint.kt
@@ -1,6 +1,5 @@
 package net.corda.finance.contracts.universal
 
-import net.corda.core.utilities.organisation
 import net.corda.core.crypto.toStringShort
 import net.corda.core.identity.Party
 import java.math.BigDecimal
