+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -7,7 +7,6 @@ import net.corda.core.internal.*
 import net.corda.core.internal.concurrent.thenMatch
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.utilities.loggerFor
-import net.corda.core.utilities.organisation
 import net.corda.node.*
 import net.corda.node.services.config.FullNodeConfiguration
 import net.corda.node.services.transactions.bftSMaRtSerialFilter
