+++ node/src/main/kotlin/net/corda/node/services/keys/KMSUtils.kt
@@ -1,9 +1,9 @@
 package net.corda.node.services.keys
 
 import net.corda.core.crypto.Crypto
-import net.corda.core.utilities.cert
 import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.node.services.IdentityService
+import net.corda.core.utilities.cert
 import net.corda.core.utilities.days
 import net.corda.node.utilities.CertificateType
 import net.corda.node.utilities.ContentSignerBuilder
