+++ node/src/test/kotlin/net/corda/node/services/config/FullNodeConfigurationTest.kt
@@ -1,6 +1,5 @@
 package net.corda.node.services.config
 
-import net.corda.core.utilities.organisation
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.nodeapi.User
 import net.corda.testing.ALICE
