+++ tools/demobench/src/main/kotlin/net/corda/demobench/model/NodeConfig.kt
@@ -2,9 +2,7 @@ package net.corda.demobench.model
 
 import com.typesafe.config.*
 import net.corda.core.identity.CordaX500Name
-import net.corda.core.utilities.locality
 import net.corda.nodeapi.User
-import org.bouncycastle.asn1.x500.X500Name
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
