+++ tools/explorer/src/main/kotlin/net/corda/explorer/formatters/PartyNameFormatter.kt
@@ -1,8 +1,6 @@
 package net.corda.explorer.formatters
 
 import net.corda.core.identity.CordaX500Name
-import net.corda.core.utilities.organisation
-import org.bouncycastle.asn1.x500.X500Name
 
 object PartyNameFormatter {
     val short = object : Formatter<CordaX500Name> {
